% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSynth.prep.R
\name{tsSynth.prep}
\alias{tsSynth.prep}
\title{Construct a list of matrices from panel dataset to be loaded into other functions needed.}
\usage{
tsSynth.prep(
  data = NULL,
  treatment.identifier = NULL,
  controls.identifier = NULL,
  time.variable = NULL,
  time.prior = NULL,
  time.post = NULL
)
}
\arguments{
\item{data}{The dataframe with required wide format.}

\item{treatment.identifier}{A scalar identifying the column number of treatment unit in the dataframe or a character string giving the column name of treatment unit in the dataframe.}

\item{time.variable}{A scalar identifying column number or column-name character string associated with period (time) data. The time variable has to be numeric.}

\item{time.prior}{A numeric vector identifying the row numbers corresponding to pre-treatment periods.}

\item{time.post}{A numeric vector identifying the row numbers corresponding to post-treatment periods.}

\item{control.identifier}{A scalar identifying the column numbers of control units in the dataframe or a vector of character strings giving the column names of control units in the dataframe.}
}
\value{
A list with a series of components prepared for running \code{tsSynth.meth}.
\item{Y1}{A matrix of treatment unit data for pre-treatment periods.}
\item{X1}{A matrix of control units data for pre-treatment periods.}
\item{Y2}{A matrix of treatment unit data for post-treatment periods.}
\item{X2}{A matrix of control units data for post-treatment periods.}
\item{time}{A matrix of period(time) data.}
\item{T}{A scalar identifying the column number of period(time) data.}
\item{T1}{A scalar identifying the number of pre-treatment periods.}
\item{T2}{A scalar identifying the number of post-treatment periods.}
\item{n}{A scalar identifying the column number of control units data.}
}
\description{
This function extracts relevant data objects from the given dataframe and produce a list of matrices necessary for running appropriate synthetic control method.
}
\details{
Users should import a dataframe ("data") with required wide format, identify the columns associated with treatment and control units respectively, time variable, the pre-treatment and post-treatment time periods.
}
\examples{
  ##First example: wide-format toy dataset sourced from Package "Synth".

  #load data from the package.
  data(synth.data)

  #extract relevant components necessary for running tsSynth.meth()
  #from wide-format panel data.
  tsSynth.prep.out =
    tsSynth.prep(
     synth.data,
     treatment.identifier = 9,
     controls.identifier = c(2:8),
     time.variable = 1,
     time.prior = c(1984:1989),
     time.post = c(1990:1998)
     )

  ## Second example: The economic impact of terrorism in the
  ## Basque country using data from Abadie and Gardeazabal (2003)

  #load data from the package.
  data(basque)

  #extract relevant components necessary for running tsSynth.meth()
  #from wide-format panel data.
  tsSynth.prep.out2 =
    tsSynth.prep(
     data = basque,
     treatment.identifier = 5,
     controls.identifier = c(2:4,6:19),
     time.variable = "year",
     time.prior = c(1964:1969),
     time.post = c(1970:1985)
     )

}
\references{
Li K, Shankar V (2020) Estimating the causal effect of a digitally native retailer opening a new store: a new two-step synthetic control method. Working Paper, University of Texas at Austin, Austin, TX.

  Li K (2019) Statistical inference for average treatment effects estimated by synthetic control methods. \emph{Journal of American Statistical Association}.
}
